Add ExposureMode
Note that there is no way in the current class to identify something without a default as "Not set" (like ExposureMode).

I initialize the value to std::numeric_limits<unsigned short>::max() as something that can be checked to see if it was set or not in the EXIF data.